feat(mergeInMap): add deep merge functionality to mergeInMap and delete deepMergeInMap
